requirements.txt
#BATCHING docker requirements, which is those that getting installed in the docker image (not all is used in jupyter notebook)
python-dotenv
entsoe-py==0.5.9
pyspark==3.3.1
pandas
sqlalchemy 
psycopg2-binary
beautifulsoup4
apache-airflow-providers-google
apache-airflow-providers-apache-spark
apache-airflow-providers-dbt-cloud
apache-airflow-providers-github
apache-airflow-providers-cncf-kubernetes
google-auth
google-cloud
google-cloud-bigquery
google-cloud-core
google-cloud-storage
pytz
xmltodict
pytest
pytest-mock


